+++ src/main/scala/ztis/model/ModelBuilder.scala
@@ -54,7 +54,8 @@ object ModelBuilder extends App with StrictLogging {
         s"mkdir $modelDirectory" !!
 
         val model = ALS.train(training, rank, numIter, lambda)
-        val score = evaluate(model, modelDirectory)
+        val predictor = new ALSPrediction(model)
+        val score = Evaluations.evaluateAndGiveAUC(predictor, validation, modelDirectory)
 
         (score, model, params)
       }
@@ -68,9 +69,11 @@ object ModelBuilder extends App with StrictLogging {
     persistInDatabase(model)
     logger.info(s"The best model params: rank=$rank, lambda=$lambda, numIter=$numIter. It's ROC AUC = $score")
 
+    val noPersonalizationPredictor = new NoPersonalizationPrediction(training, defaultRank = 2.5)
+
     s"mkdir ${directory + "/baseline-product-mean"}" !!
 
-    compareWithNoPersonalizationPrediction(directory + "/baseline-product-mean")
+    Evaluations.evaluateAndGiveAUC(noPersonalizationPredictor, validation, directory + "/baseline-product-mean")
   }
 
   spark.stop()
@@ -82,132 +85,6 @@ object ModelBuilder extends App with StrictLogging {
     }.mkString("\n")
 
     File(filename).writeAll(header, csv)
-  }
-
-  def evaluate(model: MatrixFactorizationModel, reportDirectory: String) = {
-    val userProducts = validation.map(rating => (rating.user, rating.product))
-    val predictions = model.predict(userProducts)
-    val keyedPredictions = predictions.map(rating => ((rating.user, rating.product), rating.rating))
-    val keyedObservations = validation.map(rating => ((rating.user, rating.product), rating.rating))
-    val predictionAndObservations = keyedPredictions.join(keyedObservations).values
-
-    val regressionMetrics = new RegressionMetrics(predictionAndObservations)
-    val rmse = regressionMetrics.rootMeanSquaredError
-
-    def isGoodEnough(observation: Double) = if (observation > 3) 1.0 else 0.0
-
-    val scoresAndLabels = predictionAndObservations.map {
-      case (prediction, observation) => (prediction, isGoodEnough(observation))
-    }
-
-    scoresAndLabels.cache()
-    val fraction = 1000.0 / scoresAndLabels.count()
-
-    val metrics = new BinaryClassificationMetrics(scoresAndLabels)
-    val prArea = metrics.areaUnderPR()
-    val rocArea = metrics.areaUnderROC()
-
-    shortRddToFile(reportDirectory + "/pr.dat",
-      metrics.pr().map {
-        case (recall, precision) => s"$recall, $precision"
-      }.sample(false, fraction)
-    )
-    shortRddToFile(reportDirectory + "/roc.dat",
-      metrics.roc().map {
-        case (rate1, rate2) => s"$rate1, $rate2"
-      }.sample(false, fraction)
-    )
-
-    shortRddToFile(reportDirectory + "/thresholds.dat",
-      metrics.precisionByThreshold().join(metrics.recallByThreshold()).join(metrics.fMeasureByThreshold()).map {
-        case (threshold, ((precision, recall), fMeasure)) => s"$threshold, $precision, $recall, $fMeasure"
-      }.sample(false, fraction)
-    )
-
-    val report =
-      s"""|
-          |RMSE: $rmse
-          |AUC PR: $prArea
-          |AUC ROC: $rocArea
-          |""".stripMargin
-
-    logger.info(report)
-    File(reportDirectory + "/report.txt").writeAll(report)
-
-    s"./plots.sh $reportDirectory" !!
-
-    rocArea
-  }
-
-  def compareWithNoPersonalizationPrediction(reportDirectory: String) = {
-    val predictor = new NoPersonalizationPrediction(training, defaultRank = 2.5)
-
-    val userProducts = validation.map(rating => (rating.user, rating.product))
-    val predictions = predictor.predictMany(userProducts)
-
-    val keyedPredictions = predictions.map(rating => ((rating.user, rating.product), rating.rating))
-    val keyedObservations = validation.map(rating => ((rating.user, rating.product), rating.rating))
-    val predictionAndObservations = keyedPredictions.join(keyedObservations).values
-
-    val regressionMetrics = new RegressionMetrics(predictionAndObservations)
-    val rmse = regressionMetrics.rootMeanSquaredError
-
-    def isGoodEnough(observation: Double) = if (observation > 3) 1.0 else 0.0
-
-    val scoresAndLabels = predictionAndObservations.map {
-      case (prediction, observation) => (prediction, isGoodEnough(observation))
-    }
-
-    val metrics = new BinaryClassificationMetrics(scoresAndLabels)
-    val prArea = metrics.areaUnderPR()
-    val rocArea = metrics.areaUnderROC()
-
-    shortRddToFile(reportDirectory + "/pr.dat",
-      sampleAtMost(1000, metrics.pr().map {
-        case (recall, precision) => s"$recall, $precision"
-      })
-    )
-    shortRddToFile(reportDirectory + "/roc.dat",
-      sampleAtMost(1000, metrics.roc().map {
-        case (rate1, rate2) => s"$rate1, $rate2"
-      })
-    )
-
-    shortRddToFile(reportDirectory + "/thresholds.dat",
-      sampleAtMost(1000,
-        metrics.precisionByThreshold().join(metrics.recallByThreshold()).join(metrics.fMeasureByThreshold()).map {
-          case (threshold, ((precision, recall), fMeasure)) => s"$threshold, $precision, $recall, $fMeasure"
-        }
-      )
-    )
-
-    val report =
-      s"""|
-         |RMSE: $rmse
-          |AUC PR: $prArea
-          |AUC ROC: $rocArea
-          |""".stripMargin
-
-    logger.info(report)
-    File(reportDirectory + "/report.txt").writeAll(report)
-
-    s"./plots.sh $reportDirectory" !!
-
-    rocArea
-  }
-
-  private def sampleAtMost(elements: Int, rdd: RDD[String]) = {
-    val count = rdd.count()
-    if(count <= elements) {
-      rdd
-    }
-    else {
-      rdd.sample(false, elements / count.toDouble)
-    }
-  }
-
-  private def shortRddToFile(filename: String, rdd: RDD[String]) = {
-    File(filename).writeAll(rdd.collect().mkString("\n"))
   }
 
   private def persistInDatabase(model: MatrixFactorizationModel) = {
